refactor: extract util from hook
